Separation of concern.
